move checkers to single interface
