+++ pkg/http/http.go
@@ -8,7 +8,6 @@ import (
 	"context"
 	"crypto/tls"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"log/slog"
@@ -19,18 +18,26 @@ import (
 	"strings"
 
 	"github.com/absmach/mgate"
-	"github.com/absmach/mgate/pkg/common"
 	"github.com/absmach/mgate/pkg/session"
 	mptls "github.com/absmach/mgate/pkg/tls"
+	"github.com/absmach/mgate/pkg/transport"
 	"github.com/gorilla/websocket"
 	"golang.org/x/sync/errgroup"
 )
 
-const contentType = "application/json"
+const (
+	contentType      = "application/json"
+	authzQueryKey    = "authorization"
+	authzHeaderKey   = "Authorization"
+	connHeaderKey    = "Connection"
+	connHeaderVal    = "upgrade"
+	upgradeHeaderKey = "Upgrade"
+	upgradeHeaderVal = "websocket"
+)
 
 func isWebSocketRequest(r *http.Request) bool {
-	return strings.EqualFold(r.Header.Get("Connection"), "Upgrade") &&
-		strings.EqualFold(r.Header.Get("Upgrade"), "websocket")
+	return strings.EqualFold(r.Header.Get(connHeaderKey), connHeaderVal) &&
+		strings.EqualFold(r.Header.Get(upgradeHeaderKey), upgradeHeaderVal)
 }
 
 func (p Proxy) getUserPass(r *http.Request) (string, string) {
@@ -38,25 +45,25 @@ func (p Proxy) getUserPass(r *http.Request) (string, string) {
 	switch {
 	case ok:
 		return username, password
-	case len(r.URL.Query()["authorization"]) != 0:
-		password = r.URL.Query()["authorization"][0]
+	case len(r.URL.Query()[authzQueryKey]) != 0:
+		password = r.URL.Query()[authzQueryKey][0]
 		return username, password
-	case r.Header.Get("Authorization") != "":
-		password = r.Header.Get("Authorization")
+	case r.Header.Get(authzHeaderKey) != "":
+		password = r.Header.Get(authzHeaderKey)
 		return username, password
 	}
 	return username, password
 }
 
 func (p Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	if !strings.HasPrefix(r.URL.Path, common.AddSuffixSlash(p.config.PathPrefix+p.config.TargetPath)) {
+	if !strings.HasPrefix(r.URL.Path, transport.AddSuffixSlash(p.config.PathPrefix+p.config.TargetPath)) {
 		http.NotFound(w, r)
 		return
 	}
 
 	r.URL.Path = strings.TrimPrefix(r.URL.Path, p.config.PathPrefix)
 
-	if err := p.bypassMatcher.ShouldBypass(r); err == nil {
+	if err := p.checkers.ShouldBypass(r); err == nil {
 		p.target.ServeHTTP(w, r)
 		return
 	}
@@ -102,123 +109,9 @@ func (p Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	p.target.ServeHTTP(w, r)
 }
 
-func (p *Proxy) handleWebSocket(w http.ResponseWriter, r *http.Request, s *session.Session) {
-	topic := r.URL.Path
-	ctx := session.NewContext(context.Background(), s)
-	if err := p.session.AuthConnect(ctx); err != nil {
-		http.Error(w, err.Error(), http.StatusUnauthorized)
-		return
-	}
-	if err := p.session.AuthSubscribe(ctx, &[]string{topic}); err != nil {
-		http.Error(w, err.Error(), http.StatusUnauthorized)
-		return
-	}
-	if err := p.session.Subscribe(ctx, &[]string{topic}); err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	header := http.Header{}
-
-	if auth := r.Header.Get("Authorization"); auth != "" {
-		header.Set("Authorization", auth)
-	}
-
-	target := fmt.Sprintf("%s://%s:%s%s", wsScheme(p.config.TargetProtocol), p.config.TargetHost, p.config.TargetPort, r.URL.RequestURI())
-
-	targetConn, _, err := websocket.DefaultDialer.Dial(target, header)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadGateway)
-		return
-	}
-	defer targetConn.Close()
-
-	inConn, err := p.wsUpgrader.Upgrade(w, r, nil)
-	if err != nil {
-		p.logger.Warn("WS Proxy failed to upgrade connection", slog.Any("error", err))
-		return
-	}
-	defer inConn.Close()
-
-	g, ctx := errgroup.WithContext(ctx)
-
-	g.Go(func() error {
-		err := p.stream(ctx, topic, inConn, targetConn, true)
-		_ = targetConn.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(websocket.CloseNormalClosure, "client closed"))
-		_ = targetConn.Close()
-		return err
-	})
-	g.Go(func() error {
-		err := p.stream(ctx, topic, targetConn, inConn, false)
-		_ = inConn.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(websocket.CloseNormalClosure, "client closed"))
-		_ = inConn.Close()
-		return err
-	})
-
-	gErr := g.Wait()
-	if err := p.session.Unsubscribe(ctx, &[]string{topic}); err != nil {
-		p.logger.Error("Unsubscribe failed", slog.String("topic", topic), slog.Any("error", err))
-	}
-	if gErr != nil {
-		p.logger.Error("WS Proxy session terminated", slog.Any("error", gErr))
-		return
-	}
-	p.logger.Info("WS Proxy session terminated")
-}
-
-func (p *Proxy) stream(ctx context.Context, topic string, src, dest *websocket.Conn, upstream bool) error {
-	for {
-		messageType, payload, err := src.ReadMessage()
-		if err != nil {
-			return handleStreamErr(err, upstream)
-		}
-		if upstream {
-			if err := p.session.AuthPublish(ctx, &topic, &payload); err != nil {
-				return err
-			}
-			if err := p.session.Publish(ctx, &topic, &payload); err != nil {
-				return err
-			}
-		}
-		if err := dest.WriteMessage(messageType, payload); err != nil {
-			return err
-		}
-	}
-}
-
-func handleStreamErr(err error, upstream bool) error {
-	if err == nil {
-		return nil
-	}
-
-	if upstream && websocket.IsCloseError(err, websocket.CloseNormalClosure, websocket.CloseNoStatusReceived) {
-		return nil
-	}
-	if errors.Is(err, net.ErrClosed) {
-		return nil
-	}
-
-	prefix := "downstream"
-	if upstream {
-		prefix = "upstream"
-	}
-	return fmt.Errorf("%s error: %w", prefix, err)
-}
-
-func checkOrigin(oc common.OriginChecker) func(r *http.Request) bool {
+func checkOrigin(oc Checkers) func(r *http.Request) bool {
 	return func(r *http.Request) bool {
 		return oc.CheckOrigin(r) == nil
-	}
-}
-
-func wsScheme(scheme string) string {
-	switch scheme {
-	case "http":
-		return "ws"
-	case "https":
-		return "wss"
-	default:
-		return scheme
 	}
 }
 
@@ -236,12 +129,12 @@ func encodeError(w http.ResponseWriter, defStatusCode int, err error) {
 
 // Proxy represents HTTP Proxy.
 type Proxy struct {
-	config        mgate.Config
-	target        *httputil.ReverseProxy
-	session       session.Handler
-	logger        *slog.Logger
-	wsUpgrader    websocket.Upgrader
-	bypassMatcher common.BypassMatcher
+	config     mgate.Config
+	target     *httputil.ReverseProxy
+	session    session.Handler
+	logger     *slog.Logger
+	wsUpgrader websocket.Upgrader
+	checkers   Checkers
 }
 
 func NewProxy(config mgate.Config, handler session.Handler, logger *slog.Logger, allowedOrigins []string, bypassPaths []string) (Proxy, error) {
@@ -250,20 +143,19 @@ func NewProxy(config mgate.Config, handler session.Handler, logger *slog.Logger,
 		Host:   net.JoinHostPort(config.TargetHost, config.TargetPort),
 	}
 
-	oc := common.NewOriginChecker(logger, allowedOrigins)
-	wsUpgrader := websocket.Upgrader{CheckOrigin: checkOrigin(oc)}
-	bypassMatcher, err := common.NewBypassMatcher(bypassPaths)
+	c, err := NewCheckers(allowedOrigins, bypassPaths)
 	if err != nil {
 		return Proxy{}, err
 	}
+	wsUpgrader := websocket.Upgrader{CheckOrigin: checkOrigin(c)}
 
 	return Proxy{
-		config:        config,
-		target:        httputil.NewSingleHostReverseProxy(targetUrl),
-		session:       handler,
-		logger:        logger,
-		wsUpgrader:    wsUpgrader,
-		bypassMatcher: bypassMatcher,
+		config:     config,
+		target:     httputil.NewSingleHostReverseProxy(targetUrl),
+		session:    handler,
+		logger:     logger,
+		wsUpgrader: wsUpgrader,
+		checkers:   c,
 	}, nil
 }
 
@@ -286,7 +178,7 @@ func (p Proxy) Listen(ctx context.Context) error {
 
 	mux := http.NewServeMux()
 
-	mux.Handle(common.AddSuffixSlash(p.config.PathPrefix), p)
+	mux.Handle(transport.AddSuffixSlash(p.config.PathPrefix), p)
 	server.Handler = mux
 
 	g.Go(func() error {
